Simplify integration tests even more
