+++ Makefile
@@ -1,6 +1,5 @@
 .DEFAULT_GOAL = build
 extension = $(patsubst windows,.exe,$(filter windows,$(1)))
-GO := go
 PKG_NAME := gomplate
 DOCKER_REPO ?= hairyhenderson/$(PKG_NAME)
 PREFIX := .
@@ -129,35 +128,35 @@ docker-images: gomplate.iid gomplate-slim.iid
 
 $(PREFIX)/bin/$(PKG_NAME)_%v5$(call extension,$(GOOS)): $(shell find $(PREFIX) -type f -name "*.go")
 	GOOS=$(shell echo $* | cut -f1 -d-) GOARCH=$(shell echo $* | cut -f2 -d- ) GOARM=5 CGO_ENABLED=0 \
-		$(GO) build \
+		go build \
 			-ldflags "-w -s $(COMMIT_FLAG) $(VERSION_FLAG)" \
 			-o $@ \
 			./cmd/$(PKG_NAME)
 
 $(PREFIX)/bin/$(PKG_NAME)_%v6$(call extension,$(GOOS)): $(shell find $(PREFIX) -type f -name "*.go")
 	GOOS=$(shell echo $* | cut -f1 -d-) GOARCH=$(shell echo $* | cut -f2 -d- ) GOARM=6 CGO_ENABLED=0 \
-		$(GO) build \
+		go build \
 			-ldflags "-w -s $(COMMIT_FLAG) $(VERSION_FLAG)" \
 			-o $@ \
 			./cmd/$(PKG_NAME)
 
 $(PREFIX)/bin/$(PKG_NAME)_%v7$(call extension,$(GOOS)): $(shell find $(PREFIX) -type f -name "*.go")
 	GOOS=$(shell echo $* | cut -f1 -d-) GOARCH=$(shell echo $* | cut -f2 -d- ) GOARM=7 CGO_ENABLED=0 \
-		$(GO) build \
+		go build \
 			-ldflags "-w -s $(COMMIT_FLAG) $(VERSION_FLAG)" \
 			-o $@ \
 			./cmd/$(PKG_NAME)
 
 $(PREFIX)/bin/$(PKG_NAME)_windows-%.exe: $(shell find $(PREFIX) -type f -name "*.go")
 	GOOS=windows GOARCH=$* GOARM= CGO_ENABLED=0 \
-		$(GO) build \
+		go build \
 			-ldflags "-w -s $(COMMIT_FLAG) $(VERSION_FLAG)" \
 			-o $@ \
 			./cmd/$(PKG_NAME)
 
 $(PREFIX)/bin/$(PKG_NAME)_%$(TARGETVARIANT)$(call extension,$(GOOS)): $(shell find $(PREFIX) -type f -name "*.go")
 	GOOS=$(shell echo $* | cut -f1 -d-) GOARCH=$(shell echo $* | cut -f2 -d- ) GOARM=$(GOARM) CGO_ENABLED=0 \
-		$(GO) build \
+		go build \
 			-ldflags "-w -s $(COMMIT_FLAG) $(VERSION_FLAG)" \
 			-o $@ \
 			./cmd/$(PKG_NAME)
@@ -169,15 +168,23 @@ build: $(PREFIX)/bin/$(PKG_NAME)_$(GOOS)-$(GOARCH)$(TARGETVARIANT)$(call extensi
 
 ifeq ($(OS),Windows_NT)
 test:
-	$(GO) test -coverprofile=c.out ./...
+	go test -coverprofile=c.out ./...
 else
 test:
-	$(GO) test -race -coverprofile=c.out ./...
+	go test -race -coverprofile=c.out ./...
 endif
 
-integration: build
-	$(GO) test -v -tags=integration \
+ifeq ($(OS),Windows_NT)
+integration: $(PREFIX)/bin/$(PKG_NAME)$(call extension,$(GOOS))
+	go test -v \
+		-ldflags "-X `go list ./internal/tests/integration`.GomplateBinPath=$(shell cygpath -ma .)/$<" \
 		./internal/tests/integration
+else
+integration: $(PREFIX)/bin/$(PKG_NAME)
+	go test -v \
+		-ldflags "-X `go list ./internal/tests/integration`.GomplateBinPath=$(shell pwd)/$<" \
+		./internal/tests/integration
+endif
 
 integration.iid: Dockerfile.integration $(PREFIX)/bin/$(PKG_NAME)_linux-amd64$(call extension,$(GOOS))
 	docker build -f $< --iidfile $@ .
